+++ static/styles/compilacao.scss
@@ -576,7 +576,7 @@ a:link:after, a:visited:after {
                                 }
                             }
                         }
-                        .norma_title {
+                        .ta_title {
                             padding: 0.15em 0.7em;
                             background-color: rgba(0, 0, 0, 0.15);
                             margin: 0.1em 0.08em 0 0.1em;
