+++ templates/compilacao/edit.html
@@ -14,7 +14,7 @@
 
 
 {% block title%}
-  <h1><b>Edição:</b> {{ view.get_norma }} - <i>{% trans 'Texto Multivigente' %}</i></h1>
+  <h1><b>Edição:</b> {{ view.get_ta }} - <i>{% trans 'Texto Multivigente' %}</i></h1>
 {% endblock %}
 
 {% block base_content %}
