+++ templates/compilacao/form.html
@@ -1,0 +1,6 @@
+{% extends "base.html" %}
+{% load i18n crispy_forms_tags %}
+
+{% block base_content %}
+  {% crispy form %}
+{% endblock %}
