+++ templates/compilacao/textoarticulado_detail.html
@@ -1,0 +1,71 @@
+{% extends "base.html" %}
+{% load i18n %}
+{% load compilacao_filters %}
+ 
+{% block base_content %}
+
+  {# FIXME is this the best markup to use? #}
+  <div class="clearfix">
+    {% block actions %}
+      <dl class="sub-nav right">
+        <dd><a href="{% url 'ta_edit' object.pk %}" class="button">{% trans 'Editar' %}</a></dd>
+        <dd><a href="" class="button alert">{% trans 'Excluir' %}</a></dd>
+      </dl>
+    {% endblock actions %}
+    {% block sections_nav %}{% endblock %}
+  </div>
+
+  {% block detail_content %}
+    {# TODO replace fieldset for something semantically correct, but with similar visual grouping style #}
+
+    <fieldset>
+      <legend>{%trans 'Identificação Básica'%}</legend>
+
+      <div class="row">
+
+          <div class="columns large-5">
+              <div id="div_id_tipo" class="holder">
+                  <label>{% verbose_name object 'tipo_ta' %}</label>
+                  <p>{{ object.tipo_ta}}</p>
+              </div>
+          </div>
+
+          <div class="columns large-2">
+              <div id="div_id_numero" class="holder">
+                  <label>{% verbose_name object 'numero' %}</label>
+                 <p>{{ object.numero}}</p>
+             </div>
+          </div>
+
+          <div class="columns large-2">
+              <div id="div_id_ano" class="holder">
+                  <label>{% verbose_name object 'ano' %}</label>
+                 <p>{{ object.ano}}</p>
+              </div>
+          </div>
+
+          <div class="columns large-3">
+              <div id="div_id_data" class="holder">
+                  <label>{% verbose_name object 'data' %}</label>
+                 <p>{{ object.data}}</p>
+              </div>
+          </div>
+      </div>
+      <div class="row">
+
+          <div class="columns large-12">
+              <div id="div_id_ementa" class="holder">
+                  <label>{% verbose_name object 'ementa' %}</label>
+                 <p>{{ object.ementa|safe}}</p>
+               </div>
+          </div>
+
+      </div>
+
+    </fieldset>
+
+
+
+  {% endblock detail_content %}
+
+{% endblock base_content %}
